requirements.txt
pip==21.3.1
python-dotenv==1.0.0
python-telegram-bot==20.0
requests==2.28.2